commit before switching branches
