+++ client/src/App.css
@@ -115,14 +115,11 @@ td {
   white-space: nowrap;
 }
 .controlPanel {
-height: 100px;
 width: 100%;
-border-radius: 20px;
-padding-left: 20px;
+border: 5px solid #f7f7f7;
+margin-top: 10px;
+white-space: nowrap;
+border-radius: 10px;
 }
-.innerControlPanel {
-  height: 100%;
-  width: 100%;
-  border-radius: 20px;
-}
+
